+++ cognite/neat/rules/models/rules/_information_rules.py
@@ -45,6 +45,7 @@ from cognite.neat.rules.models.rdfpath import (
 
 from ._base import (
     BaseMetadata,
+    DataModelType,
     ExtensionCategory,
     ExtensionCategoryType,
     MatchType,
@@ -75,6 +76,7 @@ else:
 
 class InformationMetadata(BaseMetadata):
     role: ClassVar[RoleTypes] = RoleTypes.information_architect
+    data_model_type: DataModelType = Field(DataModelType.solution, alias="dataModelType")
     schema_: SchemaCompleteness = Field(alias="schema")
     extension: ExtensionCategoryType | None = ExtensionCategory.addition
     prefix: PrefixType