+++ cognite/neat/rules/models/wrapped_entities.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+from collections.abc import Collection
 from functools import total_ordering
 from typing import Any, ClassVar, TypeVar
 
@@ -13,8 +14,7 @@ from cognite.neat.rules.models.entities import ContainerEntity, DMSNodeEntity, E
 class WrappedEntity(BaseModel, ABC):
     name: ClassVar[str]
     _inner_cls: ClassVar[type[Entity]]
-    _support_list: ClassVar[bool] = False
-    inner: Entity | list[Entity] | None
+    inner: list[Entity] | None
 
     @classmethod
     def load(cls: "type[T_WrappedEntity]", data: Any) -> "T_WrappedEntity":
@@ -38,9 +38,7 @@ class WrappedEntity(BaseModel, ABC):
         if data.casefold() == cls.name.casefold():
             return {"inner": None}
         inner = data[len(cls.name) :].removeprefix("(").removesuffix(")")
-        if cls._support_list:
-            return {"inner": [cls._inner_cls.load(entry.strip()) for entry in inner.split(",")]}
-        return {"inner": cls._inner_cls.load(inner)}
+        return {"inner": [cls._inner_cls.load(entry.strip()) for entry in inner.split(",")]}
 
     @model_serializer(when_used="unless-none", return_type=str)
     def as_str(self) -> str:
@@ -62,11 +60,7 @@ class WrappedEntity(BaseModel, ABC):
         return str(self)
 
     def as_tuple(self) -> tuple[str, ...]:
-        entities: list[str] = []
-        if isinstance(self.inner, Entity):
-            entities.append(str(self.inner))
-        elif isinstance(self.inner, list):
-            entities.extend(map(str, self.inner))
+        entities: list[str] = [str(inner) for inner in self.inner or []]
         return self.name, *entities
 
     def __lt__(self, other: object) -> bool:
@@ -91,39 +85,78 @@ T_WrappedEntity = TypeVar("T_WrappedEntity", bound=WrappedEntity)
 
 class DMSFilter(WrappedEntity):
     @abstractmethod
-    def as_filter(self, default: Any | None = None) -> dm.filters.Filter:
+    def as_dms_filter(self, default: Any | None = None) -> dm.filters.Filter:
         raise NotImplementedError
+
+    @classmethod
+    def from_dms_filter(cls, filter: dm.Filter) -> "DMSFilter":
+        dumped = filter.dump()
+        if (body := dumped.get(dm.filters.Equals._filter_name)) and (value := body.get("value")):
+            space = value.get("space")
+            external_id = value.get("externalId")
+            if space is not None and external_id is not None:
+                return NodeTypeFilter(inner=[DMSNodeEntity(space=space, externalId=external_id)])
+        elif (body := dumped.get(dm.filters.In._filter_name)) and (values := body.get("values")):
+            return NodeTypeFilter(
+                inner=[
+                    DMSNodeEntity(space=entry["space"], externalId=entry["externalId"])
+                    for entry in values
+                    if isinstance(entry, dict) and "space" in entry and "externalId" in entry
+                ]
+            )
+        elif body := dumped.get(dm.filters.HasData._filter_name):
+            return HasDataFilter(
+                inner=[
+                    ContainerEntity(space=entry["space"], externalId=entry["externalId"])
+                    for entry in body
+                    if isinstance(entry, dict) and "space" in entry and "externalId" in entry
+                ]
+            )
+
+        raise ValueError(f"Cannot convert {filter._filter_name} to {cls.__name__}")
 
 
 class NodeTypeFilter(DMSFilter):
     name: ClassVar[str] = "nodeType"
     _inner_cls: ClassVar[type[DMSNodeEntity]] = DMSNodeEntity
-    inner: DMSNodeEntity | None = None
+    inner: list[DMSNodeEntity] | None = None  # type: ignore[assignment]
 
-    def as_filter(self, default: NodeId | None = None) -> dm.Filter:
+    @property
+    def nodes(self) -> list[NodeId]:
+        return [node.as_id() for node in self.inner or []]
+
+    def as_dms_filter(self, default: Collection[NodeId] | None = None) -> dm.Filter:
         if self.inner is not None:
-            space = self.inner.space
-            external_id = self.inner.external_id
+            node_ids = [node.as_id() for node in self.inner]
         elif default is not None:
-            space = default.space
-            external_id = default.external_id
+            node_ids = list(default)
         else:
             raise ValueError("Empty nodeType filter, please provide a default node.")
-        return dm.filters.Equals(["node", "type"], {"space": space, "externalId": external_id})
+        if len(node_ids) == 1:
+            return dm.filters.Equals(
+                ["node", "type"], {"space": node_ids[0].space, "externalId": node_ids[0].external_id}
+            )
+        else:
+            return dm.filters.In(
+                ["node", "type"],
+                [
+                    {"space": node.space, "externalId": node.external_id}
+                    for node in sorted(node_ids, key=lambda node: node.as_tuple())
+                ],
+            )
 
 
 class HasDataFilter(DMSFilter):
     name: ClassVar[str] = "hasData"
     _inner_cls: ClassVar[type[ContainerEntity]] = ContainerEntity
-    _support_list: ClassVar[bool] = True
     inner: list[ContainerEntity] | None = None  # type: ignore[assignment]
 
-    def as_filter(self, default: list[ContainerId] | None = None) -> dm.Filter:
+    def as_dms_filter(self, default: Collection[ContainerId] | None = None) -> dm.Filter:
         containers: list[ContainerId]
         if self.inner:
             containers = [container.as_id() for container in self.inner]
         elif default:
-            containers = default
+            containers = list(default)
         else:
             raise ValueError("Empty hasData filter, please provide a default containers.")