+++ tests/tests_unit/rules/test_models/test_wrapped_entities.py
@@ -1,9 +1,10 @@
 from typing import Any
 
 import pytest
+from cognite.client import data_modeling as dm
 
 from cognite.neat.rules.models.entities import ContainerEntity, DMSNodeEntity
-from cognite.neat.rules.models.wrapped_entities import HasDataFilter, NodeTypeFilter, WrappedEntity
+from cognite.neat.rules.models.wrapped_entities import DMSFilter, HasDataFilter, NodeTypeFilter, WrappedEntity
 
 
 class TestWrappedEntities:
@@ -13,7 +14,7 @@ class TestWrappedEntities:
             (
                 NodeTypeFilter,
                 "NodeType(subject:person)",
-                NodeTypeFilter(inner=DMSNodeEntity(space="subject", externalId="person")),
+                NodeTypeFilter(inner=[DMSNodeEntity(space="subject", externalId="person")]),
             ),
             (
                 HasDataFilter,
@@ -32,6 +33,16 @@ class TestWrappedEntities:
                     ]
                 ),
             ),
+            (
+                NodeTypeFilter,
+                "nodeType(space:node1, space:node2)",
+                NodeTypeFilter(
+                    inner=[
+                        DMSNodeEntity(space="space", externalId="node1"),
+                        DMSNodeEntity(space="space", externalId="node2"),
+                    ]
+                ),
+            ),
         ],
     )
     def test_load(self, cls_: type[WrappedEntity], raw: Any, expected: WrappedEntity) -> None:
@@ -40,3 +51,33 @@ class TestWrappedEntities:
         assert loaded == expected
         assert repr(loaded) == repr(expected)
         assert str(loaded) == str(expected)
+
+    @pytest.mark.parametrize(
+        "filter_, expected",
+        [
+            (
+                dm.filters.HasData(containers=[dm.ContainerId(space="space", external_id="container1")]),
+                HasDataFilter(inner=[ContainerEntity(space="space", externalId="container1")]),
+            ),
+            (
+                dm.filters.Equals(["node", "type"], {"space": "space", "externalId": "node1"}),
+                NodeTypeFilter(inner=[DMSNodeEntity(space="space", externalId="node1")]),
+            ),
+            (
+                dm.filters.In(
+                    ["node", "type"],
+                    [{"space": "space", "externalId": "node1"}, {"space": "space", "externalId": "node2"}],
+                ),
+                NodeTypeFilter(
+                    inner=[
+                        DMSNodeEntity(space="space", externalId="node1"),
+                        DMSNodeEntity(space="space", externalId="node2"),
+                    ]
+                ),
+            ),
+        ],
+    )
+    def test_from_dms_filter(self, filter_: dm.Filter, expected: DMSFilter) -> None:
+        loaded = DMSFilter.from_dms_filter(filter_)
+
+        assert loaded == expected